sema: Add a string pool to deduplicate use of the same string literal multiple times.
